[opentitantool] Improve host emulation process termination
If the sub-process does not respond to SIGTERM for five seconds, the
code will attempt SIGKILL, which is sensible.

The code attempts to detect whether the repeated SIGKILL attempts have
worked, by checking if we get a ESRCH (no such process) from the
`signal()` system call.  However, it seems that doing that may not be a
reliable way of detecting termination, possibly because a process in
zombie state still counts as "existing" in the eyes of `signal()`.

This CL adapts the code to call `waitpid()` in between calls to
`signal()`, in order to check if the sub-process has terminated.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I4ec30d7bb99e56ca4c21b0390e3cec6660f3dde8