[sw,crypto] add SHA-384 implementation/test for OTBN
Adds a test for SHA-384. SHA-384 is a variant of SHA-512 with
different inital hash values and a truncated message digest.
Since the OTBN implementation only covers the base
algorithm which is identical for both SHA-512 and SHA-384
no modification of the core algorithm is needed and
only a note is added indicating that it is suitable for
both variants.

Signed-off-by: Felix Miller <felix.miller@lowrisc.org>