Avoid mutating a global in otbn_tracer.sv
If you build the previous code with a recent verilator, it complains
about updating "trace_output_buffer" with a blocking assignment in the
always block at the bottom of the file. I think it has inferred that
this block is sequential logic and shouldn't contain blocking
assignments. Grr!

This change switches from a mutated global variable to explicitly
passing the string that we want to update and returning the updated
value. The end result should be the same.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>