[aes, sw, sca] Use separate software LFSR for key masking
Previously, we were using the same software PRNG used for generating
plaintexts and keys. This was non-ideal for the following reasons:
1. One must be careful to keep the software PRNG on the target and the
   host in sync. Not only can this be difficult but it also limits
   capture performance on the host side.
2. This PRNG is not available in non-batch mode. As a result, non-batch
   mode was always using unmasked keys which led to 1st-order leakage.

Therefore, this PR uses an additional software LFSR for key masking
(similar to what is done for KMAC already) including a simple serial
command to seed that LFSR. Key re-masking can still be disabled by
seeding this LFSR with 0.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>