[aes, sw] Switch idle test to use ECB decryption in auto mode
Previously, ECB encryption was used in manuel mode which required some
quirks to allow the test triggering AES, disable the clock hint and then
observe the clock was still running.

The problem with these quirks is that they don't work reliably if the
start trigger delay (needed for SCA on FPGA) is adjusted. Switching
to ECB decryption and auto mode is a better an more reliable way for
this test.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>