[entropy_src, dv] Tune entropy_src_rng_max_rate config for coverage
This commit tunes the configuration of the max rate test to improve
functional coverage. This is achieved 1) by lowering the average time
between random re-configurations, doing more configs with FIPS mode
enabled, and fewer runs with ES_TYPE set to true (SHA3 conditioner off),
and 2) by increasing the number of runs.

This helps to push the functional coverage for the cr_config cross
of the csrng_hw_cg covergroup from roughly 70% to 96%. The remaining

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>