Don't try to work around NFS latency in dvsim.py
There was some code that tried to do `ls -d` in the scratch root, with
a 1 second timeout. The idea was that we'd silently work in a faster
location if the directory was a bit slow. It's neat, but probably a
source of confusion, and it's probably better for the user to set
their working directory explicitly.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>