[dvsim] Remove check that scratch root is writable in dvsim.py
Without this check, you now get something like this:

  $ util/dvsim/dvsim.py hw/top_earlgrey/dv/chip_sim_cfg.hjson --scratch-root=/root
  INFO: [dvsim] [proj_root]: /home/rjs/work/opentitan
  Traceback (most recent call last):
    File "/home/rjs/work/opentitan/util/dvsim/dvsim.py", line 751, in <module>
      main()
    File "/home/rjs/work/opentitan/util/dvsim/dvsim.py", line 672, in main
      (Path(args.scratch_root) / 'FUSESOC_IGNORE').touch()
    File "/usr/lib/python3.11/pathlib.py", line 1109, in touch
      fd = os.open(self, flags, mode)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^
  PermissionError: [Errno 13] Permission denied: '/root/FUSESOC_IGNORE'

which seems reasonable to me. Note that if we were really
enthusiastic, we could probably have fooled the existing
check (suppose the directory is writable, but the file we need to
create exists and is read-only). Probably no point in trying this at
all.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>