[bazel] fix default `otp_seed` attr default
The `otp_seed` attr of the `scramble_flash_vmem` build rule was not
configured to use the custom build configuration setting (also named
`otp_seed`). This was causing the `chip_sw_power_virus` and
`chip_sw_flash_scrambling_smoketest` to fail in the nightly regressions
as dvsim attempts to override the default value.

This should (finally) fix #17181.

Signed-off-by: Tim Trippel <ttrippel@google.com>