[openocd/rv_dm] Workaround rv_dm bug in OpenOCD by retrying failed commands
See issue #18087 for more details. In short, this commits works around the
problem by detecting when a DMI operations fails, wait for the DM/DMI to
get unstuck, reset the DMI and retry. It logs an ERROR so that problem does
not go unoticed.

The "DMI is stuck" condition is easily detectable because by requesting a
DMI reset, we get back the current value of DTMCS. This value is (per spec
and per implementation) always of the form 0x0000xyzt. When stuck, we will
get 0xffffffff which is clearly invalid. When that happens, we continue
issuing DMI resets until we see something different.

At the moment, and it is unclear why, we never get "partly invalid" values,
ie we either get a correct DTMCS value or 0xffffffff which is why the
detection above works. If it turns out that this assumption is wrong,
this patch might fail to detect the problem.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>