[sw/device] Further optimize CRC32
I noticed the disassembly of `crc32_add()` showed a number of
unnecessary reads and writes to the `uint32_t *ctx` parameter.

This commit effectively inlines `crc32_add8()` and `crc32_add32()` into
`crc32_add()` and uses a local variable as the context. This seems to
enable the compiler to keep the context in a register.

This brings the perf test to 28x faster than the first baseline, and
1.3x faster than the previous optimization in PR #17989.

    I00001 crc32_perftest.c:33] CRC32 computed in 11353 cycles.
    I00002 crc32_perftest.c:33] CRC32 computed in 11320 cycles.
    I00003 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00004 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00005 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00006 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00007 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00008 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00009 crc32_perftest.c:33] CRC32 computed in 11314 cycles.
    I00010 crc32_perftest.c:33] CRC32 computed in 11314 cycles.

Signed-off-by: Dan McArdle <dmcardle@opentitan.org>