[csrng, rtl] Simplify main input mux
Previously, this main entropy mux had multiple all-zero inputs that were
not actually required from a functionality perspective. The only case
where the design needs the input to be really zero is when flag0 is set.

This commit thus implements the following changes to the mux:
1. Reduce the number of all-zero inputs (lower FI attack surface). The
   only all-zero input is when the FI-hardened flag0 field is set.
2. Only flop new input from entropy source if an entropy request is
   actually outstanding. Previously, a glitch on the acknowledge line
   would have caused CSRNG to flop whatever was on the bus (it wouldn't
   necessarily have been consumed downstream though).

This resolves lowRISC/OpenTitan#16564.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>