[crypto] add RSA 3072 pub-key encryption
During provisioning, we need to generate an RMA token on-device and
encrypt it with an HSM generated pubkey before exporting it from the
device (see #17393). The current crypto lib only supports doing a signature
verification, not just a simple encryption operation. This adds a function
that can simply encrypt a 3072 bit value with a pubkey.

Signed-off-by: Tim Trippel <ttrippel@google.com>