[top-level,clkmgr] Allow higher clock counts variability
Allow a 3% cycle count variability in the various clocks, per the AST
specification. The settings were too tight, which is okay for open source
testing, but fail when using more accurate AST models.

The AON clock is also subject to 3% variability, so the effective variability
is set to 5%i, plus an extra cycle for CDC synchronizers.

Fixes #18037

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>