[otbn/dv] Fix differentiation between fatal alert and RMA request
Prior to this commit, OTBN's scoreboard would not distinguish between a
fatal alert and an RMA request when it saw the model transition to the
Locked state; it would always expect a fatal alert.  However, when OTBN
(and the model with it) become locked in reaction to an RMA request,
OTBN does not raise a fatal alert but acknowledges the RMA request.

This led to false test failures, e.g., `otbn_escalate` with seed
876289254 on Xcelium.

This commit fixes that; OTBN's scoreboard now expects an RMA acknowledge
instead of a fatal alert in reaction to an RMA request.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>