[test] Remove workaround from e2e_bootstrap_rma test
This workaround took the device out of `BOOTSTRAP_RMA` and delayed for a little
while to enable transitions into RMA. This was necessary because OTBN was not
ready for RMA while in the spin loop that the `BOOTSTRAP_RMA` strapping creates.

See <https://github.com/lowRISC/opentitan/issues/17944> for details.

With <https://github.com/lowRISC/opentitan/pull/18123>, the chip is now able to
transition to RMA while in the `BOOTSTRAP_RMA` strapping. This commit removes
the workaround.

The `CREATOR_SW_CFG_RMA_SPIN_CYCLES` value is increased because `10` cycles is
not enough to connect and transition within.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>