[entropy_src] Define ObserveFifoDepth parameter in hjson
Previously, RTL and DV had separate definitions of this parameter. In
addition, the hjson is using hard coded bit widths for the
OBSERVE_FIFO_DEPTH and OBSERVE_FIFO_THRESH CSRs. Therefore it's cleaner
to have only one parameter defintion and move that into the hjson
together with a comment saying that these bit widths always have to
be manually updated.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>