[rom] Skip `entropy_src` health config
Skip `entropy_src` health config if the
`CREATOR_SW_CFG_RNG_HEALTH_CONFIG_DIGEST` OTP word is not set. This is
to avoid configuring the health check values when the SW_CFG OTP partition
has not been configured.

This resolves an issue with being able to initialize the `ROM` in
`test_unlocked[0-7]` lc stages, when the OTP partions have not been
fully initialized.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>