[opentitantool] Add SPHINCS+ image commands
Adds CLI for updating the SPHINCS+ parts of the manifest.

This builds on top of the RSA signing flow; the commands for SPHINCS+
are the exact same.

As part of this, some of the arguments for RSA signing have been changed
in opentitnatool image manifest update.
  `--key-file` -> `--rsa-key`
  `--signature-file` -> `--rsa-signature`
  `--sign` has been dropped. The tool now interprets being passed a
  private key as an indication that the key should be used to sign the
  image.

--spx-key and --spx-signature have also been added as arguments to the
update sub-command.

Finally, the verify sub-command has received the --spx flag if you want
to perform the verification with SPHINCS+ signature verification (off by
default).

Signed-off-by: Jon Flatley <jflat@opentitan.org>