[site] Change default content width
For the Hugo-generated pages within current site, we only have a single
page (https://opentitan.org/documentation/index.html) that has a significant
amount of information within the 'content' div.
Currently, this page looks badly formatted as the text is significantly wider
than the diagrams (at least at 100% zoom on a 16:9 display).
Change the width limit to '$max-width-s' (currently 52rem), as this works
better for the current content.

If make more use of the Hugo-generated pages, we may need to reconsider how
we manage the width of content. Personally I do not know any of the nuance of
managing such considerations.

Signed-off-by: Harry Callahan <hcallahan@lowrisc.org>