[manuf] encrypt RMA unlock token with AES during provisioning
This uses the ECDH generated shared secret key to encrypt the RMA unlock
token with AES-ECB before exporting the token from the device. We use
ECB mode without any padding because the RMA unlock token is exactly
128-bits, i.e., the length of a single AES block.

A future PR will test the decryption of the token on the host side.

This partially addresses #17393.

Signed-off-by: Tim Trippel <ttrippel@google.com>