[aes, rtl] Update S-Box data/mask in sync with PRD
Previously, it was possible that DOM S-Box instances would get new data
and mask inputs without receiving new PRD / updating the internally
buffered PRD which is not ideal from an SCA perspective. Even if the
S-Box isn't evaluated this may create SCA leakage in the first
multiplier stage.

This commit modifies the design as follows:
- A new input is added to the DOM S-Box to directly control the
  internal PRD buffering. Upon writing state / key registers also
  the buffering is now updated.
- Inside the key expand module, some signals are added to inhibit or
  enforce PRD buffer updates irrespective of key register updates
  because some operation modes require this.

With these changes, the non-linear logic inside the S-Boxes is always
presented with both fresh data/masks and PRD except for the following
special cases:
- Key expand module: Switching betweeen FWD and INV mode of operation,
  e.g. for decryption key generation and upon switching configuration
  (rare).
- Data path: The very first cycle of the initial round for encryptions
  and decryptions. This is not critical as in the initial round the
  masked plaintext/ciphertext (not secret) is presented to the S-Boxes
  but the S-Boxes aren't evaluated.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>