[opentitantool] Recovery from failed HyperDebug firmware update
The Ti50 team plans a test lab, with hands-off HyperDebug devices
connected to OpenTitan as well as previous generation security chips.

As the first thing, before running a test suite, our script currently
runs: `opentitantool --interface hyperdebug transport update-firmware`,
which ensures that the firmware version of the debugger matches the
current version of opentitantool.  (opentitantool will only rarely be
updated, but it is cheap to perform the `update-firmware` as it skips if
the USB version string of HyperDebug already matches.)

We have become aware of a possible error case, if the update process is
interrupted, then HyperDebug will remain in the DFU bootloader mode,
meaning that it presents itself with alternative USB VID:DID, resulting
in `--interface hyperdebug` not recognising the device.  We do not want
to always have the backend driver accept the DFU identification, as the
debugger would not be able to do any of the usual operations.  However,
when asked to update firmware, it would be desirable that
`opentitantool` would accept either VID:DID.

Already, the special-purpose backend driver `--interface hyperdebug_dfu`
exists, and accepts only the DFU bootloader identification, this can be
used to recover a HyperDebug in DFU mode.  However, in order to keep our
scripts simple, with this PR I propose to have `hyperdebug_dfu` accept
either the normal or the DFU USB identification, allowing our script to
run: `opentitantool --interface hyperdebug_dfu transport
update-firmware` and be confident that it will work both in the happy
case, as well as when HyperDebug was previously left in DFU mode.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ic27134f5ace917bcbe2672ce7d705e3f9d27673b