[manuf] update RMA unlock token ECDH export keys
This updates the RMA unlock token ECDH export keys to encode them in the
PKCS8 format since it is easier to use with the Rust `elliptic-curve`
crate. Additionally this removes the public key DER file as it is not
needed on the Rust side, and the ECDH public key can be derrived from the
private key by OpenSSL when creating the public key header file for the
device side provisioning code.

Signed-off-by: Tim Trippel <ttrippel@google.com>