[opentitanlib] add support for reading/writing RISC-V registers via JTAG
This can handle GPR and CSR. At the moment only a subset of the registers
is supported but more can be added easily. The code is structured in a way
that we can also easily support for read/writes to arbitrary/custom abstract
DM  registers by adding more items to the enumeration.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>