[opentitanlib] add vmem function to merge sections
The vmem files produced by vivado tend to have many many
contiguous and small sections. This is highly inefficient
when loading them via JTAG where we really want large
section. This method tries to merge together contiguous
section. It assumes that the sections are already sorted
by increasing addresses which is the case in the test files.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>