[dv/clk_rst_if] Avoid freeze due to rst undriven
If the clk_rst_if is not driving rst_n it is possible it will never
go active. In these cases it is safe to just start the clock without
waiting.
Add a wait for rst_n, via xbar_clk_rst_if, in the xbar testbench.
Move this wait to the end of the block since the rest of the actions
take zero time and must be started ASAP.

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>