[aon_timer,dv] Fix interrupt checking
Previously due to a missed '!' a check for a reset was firing out of
reset which caused the checking logic for the wakeup and watchdog to
terminate immediately after starting. This left significant behaviour
during test execution unchecked.

This corrects that error and then fixes issues in the checking logic
that were discovered following the initial fix:

* Watchdog checking split into seperate bark and bite checkers. This
gives correct checking behaviour when they fire close to each other
* Corrected clock intervals for waiting for enable and interrupt changes
to propagate
* Corrected checks to look at intr_vif for the actual interrupts
(aon_intr_vif supplies the wakeup and reset signals only).

Signed-off-by: Greg Chadwick <gac@lowrisc.org>