Add some missing "protected" keywords
These are for methods in classes like otbn_alu_bignum_mod_err_vseq,
which derive from otbn_intg_err_vseq. The methods are marked as
protected in the base class, but we forgot to say the same in the
subclasses and it seems that Xcelium spits out a warning to say that
we've forgotten. Silence it!

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>