[manuf] perform LC transition with decrypted RMA unlock token
This attempts to perform an LC transition with the decrypted RMA unlock
token, after provisioning it into the device in the
`provisioning_functest` to confirm the unlock token was properly
provisioned, encrypted, transmitted over the console, and decrypted.

Additionally, this fixes a bug in the device-side code: the encrypted
RMA unlock token needed to be saved to the non-volatile flash region to
enable resetting the device before transmitting it over the console so
it could be committed in the SECRET2 OTP partition before attempting to
use it to transition to RMA.

This fixes #17393.

Signed-off-by: Tim Trippel <ttrippel@google.com>