[ottlib/openocd] properly handle openocd startup and stop
The current code immediately tries to connect to OpenOCD but
usually OpenOCD will make sure to have established a JTAG
connection before allowing TCL connection. This creates a
race condition that is currently now handled.
This commit fixes that by montoring the output of OpenOCD
to wait for the message saying that it is ready.
This commit also properly drops the tcp stream when stopping
OpenOCD.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>