[sigverify] Update SPHINCS+ to use aligned signatures and keys.
Since ROM public keys and signatures are already required to be
word-aligned, we can align a bunch of internal buffers in SPHINCS+ and
remove a cast-to-bytes for sigverify. The performance impact is
basically nil (about 1000 cycles faster out of a total ~1M cycles),
because I think the compiler already noticed that most of these were
aligned.

Signed-off-by: Jade Philipoom <jadep@google.com>