[spi_device] Split off early dp and cmd_info selection from outputs
The datapath selection signal had two different timing modes multiplexed
on top of a single signal, one where it would depend on flops that
change on posedge SCK and another that depended directly on SPI inputs.
Because there were negedge-triggered flops whose inputs depended on that
signal, this signal represented a high risk of hold timing failures.

Split off the function that selects a datapath to trigger internal
updates on the 8th posedge (the last clock edge for commands that are
command-only). Make this a separate signal that feeds the modules that
need it, and restore the original datapath selection signal to be purely
driven from internal flops.

Signed-off-by: Alexander Williams <awill@opentitan.org>