[opentitantool] Fix intermittent SPI issues
HyperDebug supports controlling multiple SPI busses through a single USB
interface, a "select SPI bus" command is meant to be sent before any
batch of data transfers.

When a separate function for exposing SPI EEPROM functionality was
added, this call to select spi bus was accidentally omitted.  This meant
that the working depended on the correct SPI bus already being selected
"by accident", which unfortunately happened frequently enough that the
bug was not immediately apparent.

This CL adds the call similarly to how it is done for non-EEPROM SPI
communication.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I19cae36052c401665e9b4cf04281eb6a92312624