[sdc] Add draft of SPI constraints
Four sets of SPI constraints are included here, corresponding to
different modes:
1. spi_device generating responses internally for 48 MHz full-cycle
   sampling (flash mode or generic mode)
2. spi_device generating responses internally for 24 MHz half-cycle
   sampling (TPM mode, typically)
3. spi_device passthrough for 33 MHz full-cycle sampling
4. spi_host performing full-cycle sampling

All constraints are timed for SPI mode 0.
For spi_device operation with CSB from the DIO, SPI_DEV_CS_L is promoted
to a clock. This makes for awkward constraints where it is used as data
or a reset, so the paths will need to be audited to ensure the behavior
matches expectations.

Signed-off-by: Alexander Williams <awill@opentitan.org>