[e2e tests] Do not use SRST in OpenOCD reset configuration.
Andreas has identified that OpenOCD was configured to use SRST
when issue a `reset`  command. This is problematic since it resets
the entire chip, including the JTAG tap and the pinmux.
This then causes the pinmux to disable the connection between the pins
and the tap a bit later and most likely causes the JTAG issues identified
in #17729.
The real chip will not have a srst_n pin and it should not be needed because
on RISC-V, issuing a DMI reset though ndmreset will reset *everything* but
the part necessary to keep the jtag tap alive, which is exactly what we want.

Note that the GDB tests never need to reset the entire chip, this is only needed
for LC transitions which are done in separate tests using opentitanlib and not
GDB, and that correctly handle this detail.

Hopefully fixes #17729.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>