[fpga] Explain the CSB "clock" waveform
Add a comment explaining the unusual CSB "clock" waveform. This
constraint defines a clock on CSB so it can be used for timing analysis
when it connects to clock pins of sequential elements, such as for the
status register bits. However, the waveform has been chosen to reflect
when the actual edges may occur relative to clk_spi's nominal waveform,
so that timing analysis for CSB as data also works.

Signed-off-by: Alexander Williams <awill@opentitan.org>