[gdb coordinator] Fix race condition between console and OpenOCD
A previous commit introduce a one second delay between the start of the
console, and the start of OpenOCD, to prevent interference when opentitantool
reconfigures GPIOs as part of the init when starting the console. This commit
makes this fix more robust by replacing the fixed delay with a proper sync.
We wait until opentitantool prints a specific message indicating that it is
ready. This message occurs after the GPIO reconfiguration which ensures proper
synchronization.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>