[opentitantool] Speed up HyperDebug driver
Avoid waiting for serial timeout on every HyperDebug console command.
This is achieved by stopping the processing immediately, when the
command prompt "\n> " is detected at the trailing end of characters
received from `read()` (rather than doing another `read()` with 10ms
timeout after that, and only stopping if that came back without any
characters.)

Also, rather than purging any spurious output before sending the
command, it is now done after sending the command, using the echo of the
command just sent to recognize when to start processing.  (Betting that
the probability of seeing the same command as part of the spurious
output will be vanishing, as that would otherwise cause the rest of the
spurious output to be incorrectly treated as ligitimate command output.)

This change reduces the time of `transport init` from 4.6s to 1.1s for
the Ti50 team, which would significantly speed up automated test runs.
(Currently, the actual test logic can sometimes be faster than the
setup.)

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Iea26da31676ab5607b46b644870454b52dd4eb1a