[opentitantool] Avoid no-op HyperDebug firmware updates
The special flag `--interface hyperdebug_dfu` allows opentitantool to
recognize a HyperDebug deview whether it is in its normal mode or DFU
bootloader mode.  This is particularly useful when requesting firmware
update, as doing so can recover from previous interrupted update
attempts.

However, the logic to skip updating, in case the device is already
running the newest firmware version had been omitted from the
`hyperdebug_dfu` path, probably because the first incarnation of that
mode recognized only a board already in DFU mode, and hence special
purpose for developers or first-time users.

Now that `hyperdebug_dfu` should be used as the default for scripted
updating, we want it to recognize the common case, in which updating can
be skipped.  This PR adds logic for that.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I94495d4036efa22072e9d9c81f47538cde53a7a1