[governance] Minor consistency alignments in V2 checklist
The V2S checklist refers to the V2 coverage requirements, but
unfortunately the cited thresholds were not in sync with what
the current V2 checklist says. Therefore, we remove the redundant
thresholds in the V2S checklist so that V2 becomes the sole
source of truth.

The word `stimulus` in DV_DOC_TESTPLAN_REVIEWED is replaced with
`verification` to make the statement more generic.

Signed-off-by: Michael Schaffner <msf@opentitan.org>