[chip,dv] ignore initial out of sync between jtag input and LC
When jtag driver started before cpu is enabled,
it is possible for pinmux switch jtag input to LC while jtag
input is on going. That causes LC to miss jtag IR and returns
inital junk value.
After that, it is back to normal so disable error check
when test calls `wait_lc_ready` during cpu is disabled.

Signed-off-by: Jaedon Kim <jdonjdon@opentitan.org>