[sw/device] Improve handling of Ok status in status_report
The code only exercised the error path but it may happen that an
Ok status could get added to the stack. In this case, we still want
to make sure that the code correctly prints them and to test that
output. This commit modifies the status_report_unittest to
articially push an Ok status and verifies that it is printed correctly.
It also simplifies the printing code that was unecessarily complicated.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>