[i2c] Remove SCL glitch when FMT FIFO empties mid-transaction
Prior to this change, if the FMT FIFO doesn't yet have its next entry
from software and the transaction is not done (i.e. STOP hasn't been
issued), the i2c FSM would cause a glitch on SCL. The FSM would follow
the sequence HostHoldBitAck -> PopFmtFifo -> Idle, and the Idle state
would unconditionally cause a 0 -> 1 transition on SCL, even if the user
intended to follow up with more data. Consequently, the host and device
would fall out of alignment for the bit count.

For the purposes of an ECO, change the Idle state to only release SCL if
it is not in the middle of a transaction in host mode. A subsequent
commit may change the FSM for readability (e.g. going to Active if this
IP is still in the middle of a bus transaction and fixing up the Active
state to wait for instructions from the FMT FIFO).

Signed-off-by: Alexander Williams <awill@opentitan.org>