opentitantool: Allow instantiating emulation backend w/o executable
The ordinary was to use the emulation backend is to start
opentitansession with it, giving an executable on the command line.  The
session process can then fork the emulation sub-process, and respond to
incoming network requests by manipulating the sub-process.

The test automation used by the Ti50 team wants to query each "testbed"
as part of test suite initialization, and that involves running
`opentitantool transport query` commands, which are meant to just return
values from the configuration files.  However, as the transport backend
is instantiated before the command is dispatched, the above command
currently fails for the host emulation backend.  As the very
initialiation of the backend attempts to fork a sub-process, without
having the name of the executable to start.

This CL makes the setup routine (`get_configuration()`) recognize when
no executable is given, and skip its operation.  (Should the user
attempt to start the emulation, an error will be reported by
`spawn_process()`.)  This change allows the `opentitantool transport
query` command to complete.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I3209d4e9bfaec7bd68b85a58171e5c074faf11cf